Rename: encoder -> smv_encoder
