+++ frontends/smv_cosa2.cpp
@@ -2,7 +2,7 @@
 #include "smt-switch/smt.h"
 #include "smt-switch/boolector_factory.h"
 #include "node.h"
-#include "encoder.h"
+#include "smv_encoder.h"
 #include "smvparser.tab.h"
 
 #include "bmc.h"
